+++ Chord Provider/PDFBuild/SongElements/PDFBuild+StrumSection.swift
@@ -61,9 +61,9 @@ extension PDFStringAttribute {
     /// String attributes for a strum line
     static func strumLine(settings: AppSettings) -> PDFStringAttribute {
         [
-            .foregroundColor: settings.style.theme.foreground.nsColor,
-            .font: NSFont.monospacedSystemFont(ofSize: settings.style.fonts.text.size, weight: .regular)
-        ]
+            .foregroundColor: settings.style.theme.foregroundMedium.nsColor,
+            .font: NSFont.monospacedSystemFont(ofSize: settings.style.fonts.text.size * 0.8, weight: .regular)
+        ] + alignment(.center)
     }
 
     /// String attributes for a strum line beat
@@ -71,6 +71,6 @@ extension PDFStringAttribute {
         [
             .foregroundColor: settings.style.theme.foregroundMedium.nsColor,
             .font: NSFont.monospacedSystemFont(ofSize: settings.style.fonts.text.size * 0.8, weight: .regular)
-        ]
+        ] + alignment(.center)
     }
 }