+++ Chord Provider/Views/MainScene/EditorView/EditorView+directiveButton.swift
@@ -20,5 +20,6 @@ extension EditorView {
                 Label("\(directive.details.button)…", systemImage: directive.details.icon)
             }
         )
+        .disabled(sceneState.song.definedMetaData.contains(directive))
     }
 }